Added progress tracker and reworked IO exception handling

+++ src/steelart/alex/filemanager/FileProvider.java
@@ -7,8 +7,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 
-import org.apache.commons.io.IOUtils;
-
 /**
  * This interface is needed to request file (real or temporary) for file manager element.
  * This interface implements Closeable, so it could be used in try() auto-closeable construction.
@@ -34,26 +32,42 @@ public interface FileProvider extends Closeable {
         };
     }
 
-    public static FileProvider tmpFileForInputStream(InputStream is, String name) {
+    public static FileProvider tmpFileForInputStream(InputStream is, String name, ProgressTracker progress, long size) throws IOException {
+        File tmp = null;
         try {
-            File tmp = File.createTempFile("simple-file-manager-", name);
+            File tmpFile = tmp = File.createTempFile("simple-file-manager-", name);
             tmp.deleteOnExit();
             try (OutputStream output = new FileOutputStream(tmp)) {
-                IOUtils.copy(is, output);
-                return new FileProvider() {
+                copy(is, output, progress, size);
+                FileProvider fileProvider = new FileProvider() {
                     @Override
                     public File get() {
-                        return tmp;
+                        return tmpFile;
                     }
                     @Override
                     public void close() {
-                        tmp.delete();
+                        tmpFile.delete();
                     }
                 };
+                tmp = null;
+                return fileProvider;
             }
-        } catch (IOException e) {
-            e.printStackTrace();
-            return null;
+        } finally {
+            if (tmp != null)
+                tmp.delete();
         }
+    }
+
+    public static long copy(InputStream input, OutputStream output, ProgressTracker progress, long size) throws IOException {
+        final int DEFAULT_BUFFER_SIZE = 1024 * 4;
+        byte[] buffer = new byte[DEFAULT_BUFFER_SIZE];
+        long count = 0;
+        int n = 0;
+        while (-1 != (n = input.read(buffer))) {
+            output.write(buffer, 0, n);
+            count += n;
+            progress.currentProgress(count, size);
+        }
+        return count;
     }
 }
